+++ Shared/Samples/Augemnt reality to navigate route/AugmentRealityToNavigateRouteView.swift
@@ -15,7 +15,6 @@
 import ArcGIS
 import ArcGISToolkit
 import SwiftUI
-import CoreLocation
 import AVFoundation
 
 struct AugmentRealityToNavigateRouteView: View {
@@ -38,7 +37,7 @@ struct AugmentRealityToNavigateRouteView: View {
         return scene
     }()
     /// The elevation surface set to the base surface of the scene.
-    @State var elevationSurface: Surface = {
+    @State private var elevationSurface: Surface = {
         let elevationSurface = Surface()
         elevationSurface.navigationConstraint = .unconstrained
         elevationSurface.opacity = 0.5
@@ -75,20 +74,18 @@ struct AugmentRealityToNavigateRouteView: View {
                     )
                 }
         } else {
-            VStack {
-                WorldScaleSceneView { proxy in
-                    SceneView(scene: scene, graphicsOverlays: [graphicsOverlay])
-                }
-                .calibrationButtonAlignment(.bottomLeading)
-                .task {
-                    statusText = "Adjust calibration before starting."
-                    Task {
-                        try await locationDataSource.start()
-                        
-                        for await location in locationDataSource.locations {
-                            try await routeDataModel.routeTracker?.track(location)
-                            self.elevation = location.position.z
-                        }
+            WorldScaleSceneView { _ in
+                SceneView(scene: scene, graphicsOverlays: [graphicsOverlay])
+            }
+            .calibrationButtonAlignment(.bottomLeading)
+            .task {
+                statusText = "Adjust calibration before starting."
+                Task {
+                    try await locationDataSource.start()
+                    
+                    for await location in locationDataSource.locations {
+                        try await routeDataModel.routeTracker?.track(location)
+                        self.elevation = location.position.z
                     }
                 }
             }
@@ -126,7 +123,7 @@ struct AugmentRealityToNavigateRouteView: View {
     /// Create a graphic overlay and adds a graphic (with solid yellow 3D tube symbol)
     /// to represent the route.
     @MainActor
-    func makeRouteOverlay(for routeResult: RouteResult, using routeGraphic: Graphic) -> GraphicsOverlay {
+    private func makeRouteOverlay(for routeResult: RouteResult, using routeGraphic: Graphic) -> GraphicsOverlay {
         let graphicsOverlay = GraphicsOverlay()
         graphicsOverlay.sceneProperties.surfacePlacement = .absolute
         let strokeSymbolLayer = SolidStrokeSymbolLayer(
@@ -139,13 +136,10 @@ struct AugmentRealityToNavigateRouteView: View {
         let polylineRenderer = SimpleRenderer(symbol: polylineSymbol)
         graphicsOverlay.renderer = polylineRenderer
         
-        Task {
-            if let originalPolyline = routeResult.routes.first?.geometry {
-                addElevationToPolyline(polyline: originalPolyline) { polyline in
-                    routeGraphic.geometry = polyline
-                    graphicsOverlay.addGraphic(routeGraphic)
-                }
-            }
+        if let originalPolyline = routeResult.routes.first?.geometry {
+            let polyline = addElevationToPolyline(polyline: originalPolyline)
+            routeGraphic.geometry = polyline
+            graphicsOverlay.addGraphic(routeGraphic)
         }
         
         return graphicsOverlay
@@ -157,12 +151,7 @@ struct AugmentRealityToNavigateRouteView: View {
     /// - Parameters:
     ///   - polyline: The polyline geometry of the route.
     ///   - z: A `Double` value representing z elevation.
-    ///   - completion: A completion closure to execute after the polyline is generated with success or not.
-    func addElevationToPolyline(
-        polyline: Polyline,
-        elevation z: Double = 3,
-        completion: @escaping (Polyline?) -> Void
-    ) {
+    private func addElevationToPolyline(polyline: Polyline, elevation z: Double = 3) -> Polyline {
         if let densifiedPolyline = GeometryEngine.densify(polyline, maxSegmentLength: 0.3) as? Polyline {
             let polylinebuilder = PolylineBuilder(spatialReference: polyline.spatialReference)
             
@@ -174,9 +163,9 @@ struct AugmentRealityToNavigateRouteView: View {
                 // Put the new point 3 meters above the ground elevation.
                 polylinebuilder.add(newPoint)
             }
-            completion(polylinebuilder.toGeometry())
+            return polylinebuilder.toGeometry()
         } else {
-            completion(polyline)
+            return polyline
         }
     }
     
@@ -256,230 +245,16 @@ struct AugmentRealityToNavigateRouteView: View {
     }
 }
 
-private extension AugmentRealityToNavigateRouteView {
+extension AugmentRealityToNavigateRouteView {
     @MainActor
-    struct RoutePlannerView: View {
-        /// The view model for this sample.
-        @StateObject private var model = Model()
-        /// A Boolean value indicating whether the view is showing.
-        @Binding var isShowing: Bool
-        /// The status text displayed to the user.
-        @State private var statusText = ""
-        /// User defined action to be performed when the slider delta value changes.
-        var selectRouteAction: ((Graphic, RouteResult) -> Void)?
-        /// A Boolean value indicating whether a route stop is selected.
-        var didSelectRouteStop: Bool {
-            model.startPoint != nil || model.endPoint != nil
-        }
-        /// The error shown in the error alert.
-        @State var error: Error?
-        
-        var body: some View {
-            MapView(
-                map: model.map,
-                graphicsOverlays: model.graphicsOverlays
-            )
-            .onSingleTapGesture { _, mapPoint in
-                if model.startPoint == nil {
-                    model.startPoint = mapPoint
-                    statusText = "Tap to place destination."
-                } else if model.endPoint == nil {
-                    model.endPoint = mapPoint
-                    model.routeDataModel.routeParameters.setStops(model.makeStops())
-                    Task {
-                        let routeResult = try await model.routeDataModel.routeTask.solveRoute(
-                            using: model.routeDataModel.routeParameters
-                        )
-                        if let firstRoute = routeResult.routes.first {
-                            let routeGraphic = Graphic(geometry: firstRoute.geometry)
-                            model.routeGraphicsOverlay.addGraphic(routeGraphic)
-                            model.routeDataModel.routeResult = routeResult
-                            model.didSelectRoute = true
-                            statusText = "Tap camera to start navigation."
-                        } else {
-                            self.error = error
-                        }
-                    }
-                }
-            }
-            .locationDisplay(model.locationDisplay)
-            .overlay(alignment: .top) {
-                Text(statusText)
-                    .multilineTextAlignment(.center)
-                    .frame(maxWidth: .infinity, alignment: .center)
-                    .padding(8)
-                    .background(.regularMaterial, ignoresSafeAreaEdges: .horizontal)
-            }
-            .task {
-                statusText = "Tap to place a start point."
-            }
-            .onChange(of: model.didSelectRoute) { didSelectRoute in
-                guard didSelectRoute else { return }
-                if let onDidSelectRoute = selectRouteAction,
-                   let routeResult = model.routeDataModel.routeResult {
-                    onDidSelectRoute(model.routeGraphic, routeResult)
-                }
-            }
-            .toolbar {
-                ToolbarItemGroup(placement: .bottomBar) {
-                    Spacer()
-                    Button {
-                        isShowing = false
-                    } label: {
-                        Image(systemName: "camera")
-                            .imageScale(.large)
-                    }
-                    .disabled(!model.didSelectRoute)
-                    Spacer()
-                    Button {
-                        model.reset()
-                        statusText = "Tap to place a start point."
-                        model.didSelectRoute = false
-                    } label: {
-                        Image(systemName: "trash")
-                            .imageScale(.large)
-                    }
-                    .disabled(!didSelectRouteStop)
-                }
-            }
-            .onDisappear {
-                Task { await model.locationDataSource.stop() }
-            }
-        }
-        
-        /// Sets an action to perform when the route is selected
-        /// - Parameter action: The action to perform when the route is selected.
-        func onDidSelectRoute(
-            perform action: @escaping (Graphic, RouteResult) -> Void
-        ) -> RoutePlannerView {
-            var copy = self
-            copy.selectRouteAction = action
-            return copy
-        }
+    class RouteDataModel: ObservableObject {
+        /// The route task that solves the route using the online routing service, using API key authentication.
+        @Published var routeTask = RouteTask(url: URL(string: "https://route-api.arcgis.com/arcgis/rest/services/World/Route/NAServer/Route_World")!)
+        /// The parameters for route task to solve a route.
+        @Published var routeParameters = RouteParameters()
+        /// The route tracker.
+        @Published var routeTracker: RouteTracker?
+        /// The route result.
+        @Published var routeResult: RouteResult?
     }
-}
-
-private extension AugmentRealityToNavigateRouteView.RoutePlannerView {
-    /// A view model for this example.
-    @MainActor
-    class Model: ObservableObject {
-        /// The data model for the selected route.
-        @ObservedObject var routeDataModel = RouteDataModel()
-        /// A map with an imagery basemap style.
-        @Published var map: Map = {
-            let map = Map(basemapStyle: .arcGISImagery)
-            return map
-        }()
-        /// A binding to a Boolean value indicating whether a route is selected.
-        @Published var didSelectRoute = false
-        /// The graphics overlay for the route.
-        @Published var routeOverlay: GraphicsOverlay = {
-            let graphicsOverlay = GraphicsOverlay()
-            graphicsOverlay.sceneProperties.surfacePlacement = .absolute
-            let strokeSymbolLayer = SolidStrokeSymbolLayer(
-                width: 1,
-                color: .yellow,
-                lineStyle3D: .tube
-            )
-            let polylineSymbol = MultilayerPolylineSymbol(symbolLayers: [strokeSymbolLayer])
-            let polylineRenderer = SimpleRenderer(symbol: polylineSymbol)
-            graphicsOverlay.renderer = polylineRenderer
-            
-            return graphicsOverlay
-        }()
-        /// The data source to track device location and provide updates to route tracker.
-        let locationDataSource = SystemLocationDataSource()
-        /// The graphic (with solid yellow 3D tube symbol) to represent the route.
-        @Published var routeGraphic = Graphic()
-        /// The map's location display.
-        @Published var locationDisplay: LocationDisplay = {
-            let locationDisplay = LocationDisplay()
-            locationDisplay.autoPanMode = .recenter
-            return locationDisplay
-        }()
-        /// The graphics overlay for the stops.
-        let stopGraphicsOverlay = GraphicsOverlay()
-        /// A graphic overlay for route graphics.
-        let routeGraphicsOverlay: GraphicsOverlay = {
-            let overlay = GraphicsOverlay()
-            overlay.renderer = SimpleRenderer(
-                symbol: SimpleLineSymbol(style: .solid, color: .yellow, width: 5)
-            )
-            return overlay
-        }()
-        /// The map's graphics overlays.
-        var graphicsOverlays: [GraphicsOverlay] {
-            return [stopGraphicsOverlay, routeGraphicsOverlay]
-        }
-        /// A point representing the start of navigation.
-        var startPoint: Point? {
-            didSet {
-                let stopSymbol = PictureMarkerSymbol(image: UIImage(named: "StopA")!)
-                let startStopGraphic = Graphic(geometry: self.startPoint, symbol: stopSymbol)
-                stopGraphicsOverlay.addGraphic(startStopGraphic)
-            }
-        }
-        /// A point representing the destination of navigation.
-        var endPoint: Point? {
-            didSet {
-                let stopSymbol = PictureMarkerSymbol(image: UIImage(named: "StopB")!)
-                let endStopGraphic = Graphic(geometry: self.endPoint, symbol: stopSymbol)
-                stopGraphicsOverlay.addGraphic(endStopGraphic)
-            }
-        }
-        
-        init() {
-            // Request when-in-use location authorization.
-            let locationManager = CLLocationManager()
-            if locationManager.authorizationStatus == .notDetermined {
-                locationManager.requestWhenInUseAuthorization()
-            }
-            
-            locationDisplay.dataSource = locationDataSource
-            
-            Task {
-                try await locationDataSource.start()
-                
-                let parameters = try await routeDataModel.routeTask.makeDefaultParameters()
-                
-                if let walkMode = routeDataModel.routeTask.info.travelModes.first(where: { $0.name.contains("Walking") }) {
-                    parameters.travelMode = walkMode
-                    parameters.returnsStops = true
-                    parameters.returnsDirections = true
-                    parameters.returnsRoutes = true
-                    routeDataModel.routeParameters = parameters
-                }
-            }
-        }
-        
-        /// Creates the start and destination stops for the navigation.
-        func makeStops() -> [Stop] {
-            let stop1 = Stop(point: self.startPoint!)
-            stop1.name = "Start"
-            let stop2 = Stop(point: self.endPoint!)
-            stop2.name = "Destination"
-            return [stop1, stop2]
-        }
-        
-        /// Resets the start and destination stops for the navigation.
-        func reset() {
-            routeGraphicsOverlay.removeAllGraphics()
-            stopGraphicsOverlay.removeAllGraphics()
-            routeDataModel.routeParameters.clearStops()
-            startPoint = nil
-            endPoint = nil
-        }
-    }
-}
-
-@MainActor
-private class RouteDataModel: ObservableObject {
-    /// The route task that solves the route using the online routing service, using API key authentication.
-    @Published var routeTask = RouteTask(url: URL(string: "https://route-api.arcgis.com/arcgis/rest/services/World/Route/NAServer/Route_World")!)
-    /// The parameters for route task to solve a route.
-    @Published var routeParameters = RouteParameters()
-    /// The route tracker.
-    @Published var routeTracker: RouteTracker?
-    /// The route result.
-    @Published var routeResult: RouteResult?
 }
